+++ aquatic_udp/src/config.rs
@@ -181,10 +181,10 @@ pub struct CleaningConfig {
     /// Allow clients to use a connection token for this long (seconds)
     pub max_connection_age: u32,
     /// Remove peers who have not announced for this long (seconds)
-    pub max_peer_age: u64,
+    pub max_peer_age: u32,
     /// Remove pending scrape responses that have not been returned from swarm
     /// workers for this long (seconds)
-    pub max_pending_scrape_age: u64,
+    pub max_pending_scrape_age: u32,
 }
 
 impl Default for CleaningConfig {
